Move tests from mod -> handler
